requirements.txt
flask==3.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy
Pillow
requests
opencv-python-headless
huggingface_hub
gdown
flask-cors
